Set Daemon exit reason based on exit status
This makes it possible to use the `:transient` restart mode on
Supervisors where processes get restarted only if they stop due to
errors. An exit status of 0 exits with a `:normal` reason. Any other
exit status value does not.

Previously it was only possible to use `:permanent` (restart on every
exit) and `:temporary` (never restart).