Use absolute paths for Daemon tests
It turns out that to match the System.cmd/3 symantics for Daemon, it's
necessary to raise if System.find_executable fails. Once that's done,
relative paths from the current directory only work on OSX and not
Linux. It seems like best practice to give absolute paths when you can
figure them out anyway, so do that for the tests as well.