Rewrite kill code
The child could never be killed completely without the call to wait.  It
would end up a zombie child and would be reaped by PID 1 when muontrap
exited. Since the child is reaped now, exits are faster.

Note that descendants of the child process still need to be polled since
once the child dies, they become descendants of PID 1 and wait isn't
useful.

The final change is to remove the atexit. Exiting in the atexit was
undefined and this makes it less confusing to think about getting
signals via the signal_pipe for cleanup.