Bump default wait timeout to 500 ms
Some processes just seem to take a couple hundred ms to exit. It's not
clear why some are slower than others and most are fast. It seems really
harsh to SIGKILL the slow ones that might be doing legitimate cleanup by
default. If giving up on SIGTERM quickly is important than it seems
reasonable for users to specify that.