Minor cleanup; try to stop prints from leaking out
During the tests, if the muontrap process shutdown after the Logger was
flushed, some prints could leak through the log capture. This was
harmless for the tests, but ugly for the person running the test. The
solution is to force MuonTrap.Daemon to stop before the Logger.flush().

This may need to be done in more places since it seems like logs
occasionally leak through the log capture.