Fix race where tests ran before muontrap test app was ready
The issue was that the muontrap port was being started, but it would
take a little time for muontrap to start the test program. That test
program then needed to register a signal handler, etc. On OSX, it seemed
easy to run into situations where the test app wasn't set up quickly
enough and unit tests would randomly break. While the right answer would
have been to have a handshake, given how little work needs to be done, a
quick sleep so that the OS schedules the other process seems sufficient.