Add output flow controler
There can be cases where the C process is blasting STDOUT/STDERR so
fast that it floods the MuonTrap process with messages it cannot
keep up with handling. Eventually this can build up to a state of
overloaded memory and crash the system. This is most typically seen
in smaller memory devices, like embedded systems.

This introduces a bit of flow control in the `muontrap` port to allow
pushback from the Elixir process and keep the memory buildup from
happening.

It does this by:

* Having C side having a max limit of bytes in flight before halting
  and waiting for the okay signal from Elixir
  * Elixir receives N bytes. Processes the lines and then returns
    N byte count back to the port.
  * The port adds N byte count back to the available logging limit
  * Port reads up to the logging limit worth of bytes and sends to Elixir
  * Repeat
* Changing the port processing to streaming lists of characters
  * Using `:line` port option could cause the underlying process to
    accumulate messages in the background until the line limit or
    newline character is found before formatting that and sending
    it through the `handle_info/2` callback of the GenServer
  * Since the control flow requires Elixir to mark the C process to
    allow writing, using `:line` would create situations where Elixir
    could incorrectly mark as ready while there were still lots of
    messages accumulating to be grouped by line