Add flow control to stdio
It's possible for C processes to send data on stdout and stderr so fast
that they build up in the `MuonTrap` process mailbox and cause an
out-of-memory error. This error exits Erlang. This has been seen on
embedded systems but affects all systems given enough load.

This commit adds flow control to stdio so that only 10KB (default) of
data can be in transit to Elixir at a time. The C port now intercepts
stdout and optionally stderr, and forwards data from them to Elixir
based on how much is allowed in the output window. The C port processes
acks from Elixir to allow more. When the window is closed, the C port
won't call `read(2)` on stdout/stderr which will push back on the
process that's generating the output to slow down.

The new option, `:stdio_window`, controls how much data is in flight.

Co-authored-by: Frank Hunleth <fhunleth@troodon-software.com>