Support more versions of kill to fix test failure
`kill` lists signals differently depending on the implementation. It's
possible to ask it the number for a signal and that's the most reliable
way so try it first. This works on Linux (probably all GNU inspired
versions of kill). MacOS has different opinions and this way doesn't
work there. Fall back to a second way for MacOS and systems like it.
Finally, just hardcode a number if everything fails and maybe it will be
lucky. This is just a unit test.