Label error exits as FATAL to make them more visible
The handling for exiting due to fatal errors is tricky since we need to
try to make sure the child has been killed too. These fatal errors are
rare, but we don't want to leave children running if at all possible.

This commit won't build. Error exits that can be handled were not
renamed.