Exit children when getting stdio errors
Previously, these errors were fatal exits and when they happened, the
child process wasn't cleaned up. There are a couple ways this can be
caused, but the one happening in production was traced to the Erlang
process dying which closed the port to the C muontrap process.
Simultaneously, the programming being run was generating a lot of output
(9K sent to splice(3)) which then caused splice(3) to return error since
the Erlang side closed the file descriptor. That caused the C muontrap
process to exit without killing the process it was running. That process
proceeded to do more schenanigans.