Add test to verify killing chatty processes
As the original error was timing-related, there's a possibility that
this test can produce false negatives. On my machine, the test nearly
always fails on the first iteration when the bug is present. By my
rough estimate, false negatives occured ~1-2% of the time (likely
influenced by factors like system load, which could be an issue for CI).

Closes #60.