Add a timeout option to `MuonTrap.cmd/3`
This allows callers to use `MuonTrap.cmd/3` with programs that have a
possibility of hanging forever without having to wrap the call in a
`Task`.

A child process exceeding the given timeout will be shut down in the
same way as a `MuonTrap.Daemon` (SIGTERM, followed by SIGKILL after
`:delay_to_sigkill` milliseconds). The returned value will contain the
child's output up to that point and `:timeout` in place of an exit
status.

Backward compatibility is preserved as the returned exit status will be
`:timeout` if and only if the new timeout option is present.