Add Logger metadata in `MuonTrap.Daemon`
This adds the `:muontrap_cmd` and `:muontrap_args` keys to Logger
metadata in `MuonTrap.Daemon` processes. Consumers can add additional
custom metadata to the process via the `:logger_metadata` option.

This allows for easier identification of the source of a log message
when multiple `MuonTrap.Daemon` processes are running, particularly if
they are supervising the same OS process.