Allow custom logger function in `MuonTrap.Daemon`
This enables finer-grained control over the use of the Elixir logger as
well as the option to use something other than the Elixir logger.

In an application where MuonTrap is used to supervise multiple programs,
it is now possible to differentiate between programs based on the
application and module in the log metadata. It is also now possible to
use features like compile-time log purging of MuonTrap logs on a
case-by-case basis.